+++ packages/hardhat-core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # hardhat
 
+## 2.24.0
+
+### Minor Changes
+
+- 82feaae: Set prague as the default hardfork in Hardhat network
+
+### Patch Changes
+
+- a7aa6d6: Upgraded EDR to (https://github.com/NomicFoundation/edr/releases/tag/%40nomicfoundation%2Fedr%400.11.0):
+  - Replaced const enums with non-const enums in \*.d.ts files
+- 2ab8103: Relax validations for transaction signing introduced in the previous version by disabling strict mode in `Transaction.prepare`.
+- 67f1e95: Support chainId values above 2^32 - 1 for local account transactions
+
+## 2.23.0
+
+### Minor Changes
+
+- aa09462: Add support for the Prague hardfork
+
+### Patch Changes
+
+- a5419a7: Use `npm_config_user_agent` to determine what package manager to use for project creation
+- 4ad0bc0: Upgraded EDR to [v0.9.0](https://github.com/NomicFoundation/edr/releases/tag/%40nomicfoundation%2Fedr%400.9.0):
+  - Added support for the Prague hardfork
+- 9d6f713: Upgraded EDR to [v0.10.0](https://github.com/NomicFoundation/edr/releases/tag/%40nomicfoundation%2Fedr%400.10.0):
+  - Add Prague hardfork to the list of supported `SpecId`s
+- b61cdd8: Add hardfork activation history for avalanche network
+- 2f27ee8: Display configurable banner on project creation and task execution
+
 ## 2.22.19
 
 ### Patch Changes